include option to add network with different rpc urls
