+++ stores/index.js
@@ -34,3 +34,8 @@ export const useRpcStore = create((set) => ({
   rpcs: [],
   addRpc: (value) => set((state) => ({ rpcs: [...state.rpcs, value] })),
 }));
+
+export const useAccount = create((set) => ({
+  account: null,
+  setAccount: (account) => set(() => ({ account })),
+}));
