+++ utils/utils.js
@@ -1,5 +1,7 @@
 import BigNumber from 'bignumber.js';
 import { useState, useEffect } from 'react';
+import stores from '../stores';
+import { ERROR, TRY_CONNECT_WALLET } from '../stores/constants';
 
 // todo: get navigator declared somehow? probably an issue with using nextjs
 // function getLang() {
@@ -68,3 +70,56 @@ export function useDebounce(value, delay) {
 }
 
 export const fetcher = (...args) => fetch(...args).then((res) => res.json());
+
+export const renderProviderText = (account) => {
+  if (account && account.address) {
+    const providerTextList = {
+      Metamask: 'Add to Metamask',
+      imToken: 'Add to imToken',
+      Wallet: 'Add to Wallet',
+    };
+    return providerTextList[getProvider()];
+  } else {
+    return 'Connect wallet';
+  }
+};
+
+const toHex = (num) => {
+  return '0x' + num.toString(16);
+};
+
+export const addToNetwork = (account, chain, rpc) => {
+  if (!(account && account.address)) {
+    stores.dispatcher.dispatch({ type: TRY_CONNECT_WALLET });
+    return;
+  }
+
+  const params = {
+    chainId: toHex(chain.chainId), // A 0x-prefixed hexadecimal string
+    chainName: chain.name,
+    nativeCurrency: {
+      name: chain.nativeCurrency.name,
+      symbol: chain.nativeCurrency.symbol, // 2-6 characters long
+      decimals: chain.nativeCurrency.decimals,
+    },
+    rpcUrls: rpc ? [rpc] : chain.rpc,
+    blockExplorerUrls: [
+      chain.explorers && chain.explorers.length > 0 && chain.explorers[0].url ? chain.explorers[0].url : chain.infoURL,
+    ],
+  };
+
+  window.web3.eth.getAccounts((error, accounts) => {
+    window.ethereum
+      .request({
+        method: 'wallet_addEthereumChain',
+        params: [params, accounts[0]],
+      })
+      .then((result) => {
+        console.log(result);
+      })
+      .catch((error) => {
+        stores.emitter.emit(ERROR, error.message ? error.message : error);
+        console.log(error);
+      });
+  });
+};
